BondFlow/experiment/analysis/requirements.txt
# ==========================
# 基础环境
# ==========================
numpy>=1.25
pandas>=2.1
matplotlib>=3.8
seaborn>=0.13
scipy>=1.11
scikit-learn>=1.5

# ==========================
# 并行与系统交互
# ==========================
tqdm>=4.66
joblib>=1.4
psutil>=5.9

# ==========================
# 科学计算/结构分析
# ==========================
biopython>=1.83
torch>=2.2

# ==========================
# PyRosetta 安装器（自动下载对应版本）
# ==========================
pyrosetta-installer

# ==========================
# 其他工具
# ==========================
pyyaml>=6.0